BUG: format label for inf in cut with a str directly
TST: disable sparse test on pow on win32

TST: skip msgpack testing in 2.6 for datetimes with tz

TST: dtype test compat in io/tests/test_excel.py

+++ pandas/tools/tests/test_tile.py
@@ -116,7 +116,7 @@ class TestCut(unittest.TestCase):
 
     def test_inf_handling(self):
         data = np.arange(6)
-        data_ser = Series(data)
+        data_ser = Series(data,dtype='int64')
 
         result = cut(data, [-np.inf, 2, 4, np.inf])
         result_ser = cut(data_ser, [-np.inf, 2, 4, np.inf])
